feat: first story over work.container.tsx including storybook configuration
