+++ .prettierrc
@@ -1,6 +1,6 @@
 {
   "semi": true,
   "singleQuote": true,
-  "tabWidth": 2,
+  "tabWidth": 4,
   "trailingComma": "es5"
 }
