+++ src/containers/work/work.stories.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import WorkContainer from './work.container';
+import { storiesOf } from '@storybook/react';
+
+const story = storiesOf('Work', module);
+
+story.add('Container', () => <WorkContainer />);
