--- src/theme2.js
@@ -1,30 +1,0 @@
-export const theme2 = {
-  color: {
-    yellow: '#ffff4c',
-    green: '#ACECA1',
-    orange: '#FBB02D',
-    pink: '#fe65b7',
-    blue: '#446DF6',
-    purple: '#AB87FF',
-    light: '#EFEFEF',
-    medium: '#CFCFCF',
-    dark: '#333333',
-  },
-  elementSizes: {
-    na: 0,
-    xs: 8,
-    sm: 16,
-    md: 24,
-    lg: 32,
-    xl: 40,
-  },
-  typographySizes: {
-    h1: 100,
-    h2: 80,
-    h3: 35,
-    h4: 20,
-    p: 12,
-    a: 12,
-    li: 12,
-  },
-};
