+++ tsconfig.json
@@ -9,8 +9,9 @@
     "isolatedModules": true,
     "esModuleInterop": true,
     "noUnusedLocals": false,
-    "baseUrl": "src"
+    "baseUrl": "src",
+    "rootDirs": ["src", ".storybook"]
   },
   "exclude": ["node_modules", "public", ".cache"],
-  "include": ["src"]
+  "include": ["src", ".storybook"]
 }
